style: Icons
Work done on the custom icons and styling

UIOA-143
