+++ package.json
@@ -74,11 +74,14 @@
     ]
   },
   "devDependencies": {
+    "@babel/eslint-parser": "^7.15.0",
     "@babel/plugin-proposal-class-properties": "^7.12.1",
     "@babel/plugin-proposal-decorators": "^7.12.1",
     "@babel/plugin-transform-runtime": "^7.12.1",
     "@babel/preset-react": "^7.12.5",
     "@folio/eslint-config-stripes": "^6.0.0",
+    "@folio/handler-stripes-registry": "^1.1.1",
+    "@folio/service-interaction": "^1.0.0",
     "@folio/stripes": "^7.0.0",
     "@folio/stripes-cli": "^2.3.0",
     "@folio/stripes-core": "^8.0.0",
@@ -87,7 +90,6 @@
     "@testing-library/jest-dom": "^5.11.5",
     "@testing-library/react": "^11.1.2",
     "@testing-library/user-event": "^12.2.2",
-    "@babel/eslint-parser": "^7.15.0",
     "babel-jest": "^26.6.3",
     "babel-plugin-module-resolver": "^4.0.0",
     "eslint": "^7.32.0",
@@ -106,13 +108,11 @@
     "react-intl": "^5.8.1",
     "react-query": "^3.9.0",
     "react-router-dom": "^5.2.0",
-    "sinon": "^9.0.2",
-    "@folio/handler-stripes-registry": "^1.1.1",
-    "@folio/service-interaction": "^1.0.0"    
+    "sinon": "^9.0.2"
   },
   "dependencies": {
-    "@folio/stripes-erm-components": "^6.0.0",
     "@folio/stripes-acq-components": "^3.1.0",
+    "@folio/stripes-erm-components": "^6.0.0",
     "@k-int/stripes-kint-components": "^2.8.2",
     "classnames": "^2.2.6",
     "compose-function": "^3.0.3",
