--- src/components/Checklist/CustomIcons/CrossFatIcon.js
@@ -1,8 +1,0 @@
-/* eslint-disable react/style-prop-object */
-const CrossFatIcon = (props) => (
-  <svg viewBox="0 0 32 32" xmlns="http://www.w3.org/2000/svg" {...props}>
-    <path d="m9.053 9.072 13.894 13.856M9.053 22.928 22.947 9.072" />
-  </svg>
-);
-
-export default CrossFatIcon;
