--- src/components/Checklist/CustomIcons/DashFatIcon.js
@@ -1,8 +1,0 @@
-/* eslint-disable react/style-prop-object */
-const DashFatIcon = (props) => (
-  <svg viewBox="0 0 32 32" xmlns="http://www.w3.org/2000/svg" {...props}>
-    <path d="M9.017 15.813h13.966" />
-  </svg>
-);
-
-export default DashFatIcon;
